# Use a trie with a rune slice