This needs some cleanup and some more tests, but it should be correct, and it's faster:

```
name                 old time/op    new time/op    delta
Map/No_options-10       479µs ± 1%     396µs ± 0%  -17.38%  (p=0.029 n=4+4)
Map/To_lowercase-10     506µs ± 0%     411µs ± 0%  -18.81%  (p=0.029 n=4+4)
ChristmasCarrol-10     2.78ms ± 1%    1.58ms ± 0%  -43.19%  (p=0.029 n=4+4)

name                 old alloc/op   new alloc/op   delta
Map/No_options-10       160kB ± 0%     160kB ± 0%     ~     (p=1.000 n=4+4)
Map/To_lowercase-10     160kB ± 0%     160kB ± 0%     ~     (all equal)
ChristmasCarrol-10       160B ± 0%      160B ± 0%     ~     (all equal)

name                 old allocs/op  new allocs/op  delta
Map/No_options-10        11.0 ± 0%      11.0 ± 0%     ~     (all equal)
Map/To_lowercase-10      11.0 ± 0%      11.0 ± 0%     ~     (all equal)
ChristmasCarrol-10       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```