@bep 2023-05-25
Never mind, adding som better benchmarks shows that this is not a great idea:

```

name                                               old time/op    new time/op     delta
Map/New-10                                           9.13µs ± 1%    10.92µs ± 1%     +19.62%  (p=0.029 n=4+4)
Map/New_tolower-10                                   8.88µs ± 0%    10.85µs ± 7%     +22.24%  (p=0.029 n=4+4)
Map/Map-10                                           13.7µs ± 1%      8.7µs ± 1%     -36.46%  (p=0.029 n=4+4)
Map/Map_tolower-10                                   14.7µs ± 2%      9.4µs ± 1%     -36.41%  (p=0.029 n=4+4)
Map/Map_English_Yeats_1000_words_10_keywords-10      10.5µs ± 0%      8.4µs ± 1%     -20.25%  (p=0.029 n=4+4)
Map/Map_English_Yeats_1000_words_100_keywords-10     15.3µs ± 1%      9.2µs ± 1%     -39.80%  (p=0.029 n=4+4)
Map/Map_English_Yeats_10000_words_10_keywords-10      105µs ± 0%       82µs ± 0%     -21.58%  (p=0.029 n=4+4)
Map/Map_English_Yeats_10000_words_100_keywords-10     156µs ± 1%       94µs ± 2%     -39.59%  (p=0.029 n=4+4)
Map/Map_Runes_Yeats_1000_words_10_keywords-10        8.73µs ± 0%     8.04µs ± 2%      -7.90%  (p=0.029 n=4+4)
Map/Map_Runes_Yeats_1000_words_100_keywords-10       11.5µs ± 1%      8.9µs ± 3%     -23.09%  (p=0.029 n=4+4)
Map/Map_Runes_Yeats_10000_words_10_keywords-10       90.2µs ± 1%     81.9µs ± 1%      -9.16%  (p=0.029 n=4+4)
Map/Map_Runes_Yeats_10000_words_100_keywords-10       120µs ± 1%       88µs ± 3%     -26.46%  (p=0.029 n=4+4)
Map/New_English_Yeats_1000_words_10_keywords-10      3.50µs ± 2%     4.81µs ±16%     +37.32%  (p=0.029 n=4+4)
Map/New_English_Yeats_1000_words_100_keywords-10     23.2µs ± 3%     27.1µs ± 1%     +16.58%  (p=0.029 n=4+4)
Map/New_English_Yeats_10000_words_10_keywords-10     3.44µs ± 1%     3.10µs ± 2%     -10.09%  (p=0.029 n=4+4)
Map/New_English_Yeats_10000_words_100_keywords-10    21.8µs ± 0%     27.6µs ± 0%     +26.79%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_1000_words_10_keywords-10        3.07µs ± 1%   119.36µs ± 1%   +3781.94%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_1000_words_100_keywords-10       22.7µs ± 1%    537.3µs ± 9%   +2268.02%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_10000_words_10_keywords-10       3.34µs ± 1%    70.38µs ± 3%   +2010.23%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_10000_words_100_keywords-10      20.2µs ± 0%    439.1µs ± 2%   +2076.43%  (p=0.029 n=4+4)
ChristmasCarrol-10                                   2.79ms ± 1%     1.60ms ± 0%     -42.62%  (p=0.029 n=4+4)

name                                               old alloc/op   new alloc/op    delta
Map/New-10                                           15.9kB ± 0%     78.8kB ± 0%    +394.30%  (p=0.029 n=4+4)
Map/New_tolower-10                                   14.8kB ± 0%     71.7kB ± 0%    +386.09%  (p=0.029 n=4+4)
Map/Map-10                                           2.40kB ± 0%     2.40kB ± 0%        ~     (all equal)
Map/Map_tolower-10                                   2.40kB ± 0%     2.40kB ± 0%        ~     (all equal)
Map/Map_English_Yeats_1000_words_10_keywords-10      1.12kB ± 0%     1.12kB ± 0%        ~     (all equal)
Map/Map_English_Yeats_1000_words_100_keywords-10     5.09kB ± 0%     5.09kB ± 0%        ~     (all equal)
Map/Map_English_Yeats_10000_words_10_keywords-10     10.5kB ± 0%     10.5kB ± 0%        ~     (all equal)
Map/Map_English_Yeats_10000_words_100_keywords-10    61.7kB ± 0%     61.7kB ± 0%        ~     (all equal)
Map/Map_Runes_Yeats_1000_words_10_keywords-10          160B ± 0%       160B ± 0%        ~     (all equal)
Map/Map_Runes_Yeats_1000_words_100_keywords-10       2.40kB ± 0%     2.40kB ± 0%        ~     (all equal)
Map/Map_Runes_Yeats_10000_words_10_keywords-10       2.40kB ± 0%     2.40kB ± 0%        ~     (all equal)
Map/Map_Runes_Yeats_10000_words_100_keywords-10      20.7kB ± 0%     20.7kB ± 0%        ~     (all equal)
Map/New_English_Yeats_1000_words_10_keywords-10      6.50kB ± 0%    29.46kB ± 0%    +353.57%  (p=0.029 n=4+4)
Map/New_English_Yeats_1000_words_100_keywords-10     35.8kB ± 0%    181.1kB ± 0%    +405.92%  (p=0.029 n=4+4)
Map/New_English_Yeats_10000_words_10_keywords-10     6.50kB ± 0%    21.61kB ± 0%    +232.64%  (p=0.029 n=4+4)
Map/New_English_Yeats_10000_words_100_keywords-10    34.1kB ± 0%    191.2kB ± 0%    +460.55%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_1000_words_10_keywords-10        5.50kB ± 0%  1231.43kB ± 0%  +22273.28%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_1000_words_100_keywords-10       33.0kB ± 0%   7443.2kB ± 0%  +22433.73%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_10000_words_10_keywords-10       5.99kB ± 0%  1181.26kB ± 0%  +19633.67%  (p=0.029 n=4+4)
Map/New_Runes_Yeats_10000_words_100_keywords-10      28.1kB ± 0%   6310.0kB ± 0%  +22378.40%  (p=0.029 n=4+4)
ChristmasCarrol-10                                     160B ± 0%       160B ± 0%        ~     (all equal)
```